Subsample TuningJobState for model fitting in multi-fidelity to cap its size
