+++ benchmarking/commons/hpo_main_local.py
@@ -137,6 +137,7 @@ def main(
                 random_seed=random_seed,
                 resource_attr=benchmark.resource_attr,
                 verbose=args.verbose,
+                max_size_data_for_model=args.max_size_data_for_model,
                 **method_kwargs,
             )
         )
@@ -151,6 +152,7 @@ def main(
             experiment_tag=experiment_tag,
             benchmark_name=benchmark_name,
             random_seed=master_random_seed,
+            max_size_data_for_model=args.max_size_data_for_model,
             benchmark=benchmark,
             extra_args=extra_args,
         )
