+++ benchmarking/commons/launch_remote_simulator.py
@@ -67,7 +67,7 @@ def get_hyperparameters(
         hyperparameters["start_seed"] = args.start_seed
     if args.benchmark is not None:
         hyperparameters["benchmark"] = args.benchmark
-    for k in ("n_workers", "max_wallclock_time"):
+    for k in ("n_workers", "max_wallclock_time", "max_size_data_for_model"):
         v = getattr(args, k)
         if v is not None:
             hyperparameters[k] = v
