+++ syne_tune/optimizer/schedulers/searchers/bayesopt/tuning_algorithms/defaults.py
@@ -25,3 +25,5 @@ DEFAULT_LOCAL_OPTIMIZER_CLASS = LBFGSOptimizeAcquisition
 DEFAULT_NUM_INITIAL_CANDIDATES = 250
 
 DEFAULT_NUM_INITIAL_RANDOM_EVALUATIONS = 3
+
+DEFAULT_MAX_SIZE_DATA_FOR_MODEL = 500
