+++ syne_tune/optimizer/schedulers/searchers/gp_searcher_factory.py
@@ -84,6 +84,7 @@ from syne_tune.optimizer.schedulers.searchers.bayesopt.models.meanstd_acqfunc_im
 from syne_tune.optimizer.schedulers.searchers.bayesopt.tuning_algorithms.defaults import (
     DEFAULT_NUM_INITIAL_CANDIDATES,
     DEFAULT_NUM_INITIAL_RANDOM_EVALUATIONS,
+    DEFAULT_MAX_SIZE_DATA_FOR_MODEL,
 )
 from syne_tune.optimizer.schedulers.searchers.bayesopt.datatypes.common import (
     INTERNAL_METRIC_NAME,
@@ -93,12 +94,16 @@ from syne_tune.optimizer.schedulers.searchers.bayesopt.datatypes.common import (
 from syne_tune.optimizer.schedulers.searchers.bayesopt.utils.debug_log import (
     DebugLogPrinter,
 )
+from syne_tune.optimizer.schedulers.searchers.bayesopt.utils.subsample_state import (
+    SubsampleMultiFidelityStateConverter,
+)
 from syne_tune.optimizer.schedulers.utils.simple_profiler import SimpleProfiler
 from syne_tune.optimizer.schedulers.searchers.utils.default_arguments import (
     Integer,
     Categorical,
     Boolean,
     Float,
+    IntegerOrNone,
 )
 from syne_tune.optimizer.schedulers.searchers.utils.warmstarting import (
     create_hp_ranges_for_warmstarting,
@@ -403,12 +408,17 @@ def _create_common_objects(model=None, is_hypertune=False, **kwargs):
         "skip_optimization": skip_optimization,
     }
     if is_hyperband:
+        # Extended config space
         epoch_range = (1, kwargs["max_epochs"])
         result["config_space_ext"] = ExtendedConfiguration(
             hp_ranges,
             resource_attr_key=kwargs["resource_attr"],
             resource_attr_range=epoch_range,
         )
+        # State converter to down sample data
+        max_size = kwargs.get("max_size_data_for_model")
+        if max_size is not None:
+            result["state_converter"] = SubsampleMultiFidelityStateConverter(max_size)
 
     # Create model factory
     if model == "gp_multitask":
@@ -819,6 +829,7 @@ def _common_defaults(
         default_options["separate_noise_variances"] = False
         default_options["hypertune_distribution_num_samples"] = 50
         default_options["hypertune_distribution_num_brackets"] = 1
+        default_options["max_size_data_for_model"] = DEFAULT_MAX_SIZE_DATA_FOR_MODEL
     if is_multi_output:
         default_options["initial_scoring"] = "acq_func"
         default_options["exponent_cost"] = 1.0
@@ -861,6 +872,7 @@ def _common_defaults(
         constraints["separate_noise_variances"] = Boolean()
         constraints["hypertune_distribution_num_samples"] = Integer(1, None)
         constraints["hypertune_distribution_num_brackets"] = Integer(1, None)
+        constraints["max_size_data_for_model"] = IntegerOrNone(1, None)
     if is_multi_output:
         constraints["initial_scoring"] = Categorical(choices=tuple({"acq_func"}))
         constraints["exponent_cost"] = Float(0.0, 1.0)
